Rename `Bundle` -> `ProblemSet` and `Result` -> `Answer`; solidify domain model (#88)
* Remove `bundles` components and related code

### What's done:
 * Deleted all components related to the `bundles` feature, including `BundleCard.tsx`, `BundleCategoryList.tsx`, `BundleComponent.tsx`, and others.
 * Removed associated API calls, utilities, and hooks.
 * Cleaned up redundant imports and unused code in the `bundles` directory.

* Format test files for improved readability

### What's done:
 * Reformatted JSX structure in test files for clarity and readability.
 * Applied consistent indentation and line breaks to long arguments and components.
 * Updated several test cases across multiple files, including `ProblemSetPage.test.tsx`, `ProblemListPage.test.tsx`, and others.

* Add R2DBC configuration for PostgreSQL connection

### What's done:
 * Added `r2dbc.url`, `r2dbc.username`, and `r2dbc.password` properties to `application-spec-gen.yml`.
 * Configured R2DBC to support PostgreSQL with the specified connection details.

* Add GNU AGPLv3 license to the project

### What's done:
 * Added the GNU Affero General Public License (AGPL) version 3 to the project as `LICENSE`.
 * Ensured compliance with licensing requirements by including the full license document.

* Update OpenAPI specs and add detailed endpoint descriptions

### What's done:
 * Updated OpenAPI specs for `edukate-backend`, `edukate-gateway`, and `edukate-notifier` with detailed endpoint descriptions and response schemas.
 * Renamed `OpenAPI definition` to `Edukate API` across all specs.
 * Added license (`GNU AGPLv3`) and API version (`1.0.0`) to OpenAPI configuration.
 * Enhanced request and response documentation, including summary and description fields for endpoints.
 * Included additional error response codes (`401`, `403`, `404`, etc.) for better API error reporting.

* Update ProblemSetProblemSelector dependencies

### What's done:
 * Added `onProblemSelect` and `problemSetName` to dependencies for the `useMemo` function in `ProblemSetProblemSelector.tsx`.
 * Ensured proper memoization by including all referenced variables.

* Update ProblemSetProblemSelector dependencies

### What's done:
 * Added `onProblemSelect` and `problemSetName` to dependencies for the `useMemo` function in `ProblemSetProblemSelector.tsx`.
 * Ensured proper memoization by including all referenced variables.

* Add Flyway configuration and optimize caching logic

### What's done:
 * Introduced `FlywayConfig` class to handle PostgreSQL migrations.
 * Improved caching by refining `CacheConfig` and implementing `Caching` annotations in `ProblemService`, `AnswerService`, and `UserService`.
 * Updated repository queries to support pagination without `Pageable`.
 * Replaced `setAllowedOrigins()` with `setAllowedOriginPatterns()` in CORS setup.
 * Simplified R2DBC and Flyway connection setup across configuration files.
 * Added PostgreSQL `admin` credentials in local and development environments.
 * Refactored `ProblemService` and `AnswerService` to use more precise cache invalidation logic.
 * Enhanced UI/UX in `ProblemTablePagination` and `ProblemTable`, optimizing component hierarchy.
 * Tweaked error handling in frontend for better user feedback.

* Add Flyway configuration and optimize caching logic

### What's done:
 * Introduced `FlywayConfig` class to handle PostgreSQL migrations.
 * Improved caching by refining `CacheConfig` and implementing `Caching` annotations in `ProblemService`, `AnswerService`, and `UserService`.
 * Updated repository queries to support pagination without `Pageable`.
 * Replaced `setAllowedOrigins()` with `setAllowedOriginPatterns()` in CORS setup.
 * Simplified R2DBC and Flyway connection setup across configuration files.
 * Added PostgreSQL `admin` credentials in local and development environments.
 * Refactored `ProblemService` and `AnswerService` to use more precise cache invalidation logic.
 * Enhanced UI/UX in `ProblemTablePagination` and `ProblemTable`, optimizing component hierarchy.
 * Tweaked error handling in frontend for better user feedback.

* Add `@Profile` to FlywayConfig and simplify spec-gen setup

### What's done:
 * Added `@Profile("!spec-gen")` to `FlywayConfig` to exclude it during spec generation.
 * Removed Flyway connection details from `application-spec-gen.properties`.
 * Added R2DBC connection pool configuration with `initial-size=0` for spec generation.

+++ edukate-storage/src/main/kotlin/io/github/sanyavertolet/edukate/storage/utils/SegmentFileKeyParser.kt
@@ -26,7 +26,7 @@ object SegmentFileKeyParser {
 
         // users/{userId}/tmp/{fileName}
         if (segments.size == TMP_SEGMENT_COUNT && segments[0] == "users" && segments[PATH_TYPE_INDEX] == "tmp") {
-            return TempFileKey(segments[USER_ID_INDEX], segments[TMP_FILE_NAME_INDEX])
+            return TempFileKey(segments[USER_ID_INDEX].toLong(), segments[TMP_FILE_NAME_INDEX])
         }
 
         // users/{userId}/submissions/{problemId}/{submissionId}/{fileName}
@@ -34,21 +34,21 @@ object SegmentFileKeyParser {
             segments.size == SUBMISSION_SEGMENT_COUNT && segments[0] == "users" && segments[PATH_TYPE_INDEX] == "submissions"
         ) {
             return SubmissionFileKey(
-                segments[USER_ID_INDEX],
-                segments[SUBMISSION_PROBLEM_ID_INDEX],
-                segments[SUBMISSION_SUBMISSION_ID_INDEX],
+                segments[USER_ID_INDEX].toLong(),
+                segments[SUBMISSION_PROBLEM_ID_INDEX].toLong(),
+                segments[SUBMISSION_SUBMISSION_ID_INDEX].toLong(),
                 segments[SUBMISSION_FILE_NAME_INDEX],
             )
         }
 
         // problems/{problemId}/{fileName}
         if (segments.size == SIMPLE_SEGMENT_COUNT && segments[0] == "problems") {
-            return ProblemFileKey(segments[1], segments[2])
+            return ProblemFileKey(segments[1].toLong(), segments[2])
         }
 
         // results/{problemId}/{fileName}
         if (segments.size == SIMPLE_SEGMENT_COUNT && segments[0] == "results") {
-            return ResultFileKey(segments[1], segments[2])
+            return ResultFileKey(segments[1].toLong(), segments[2])
         }
 
         throw IllegalArgumentException("Invalid key format: '$rawKey' (normalized: '${segments.joinToString("/")}')")
